% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCodes.R
\name{bss.dt.bs}
\alias{bss.dt.bs}
\title{Bayesian sample size in a decision-theoretic approach for the Birbaum-Saunders/inverse-gamma model.}
\usage{
bss.dt.bs(loss = "L1", a1 = 1E-2, b1 = 1E-2, a2 = 1E-2, b2 = 1E-2, c, rho = NULL, gam = NULL,
                nmax = 1E2, nlag = 1E1, nrep = 1E2, lrep = 1E2,
                npost = 1E2, plot = FALSE, ...)
}
\arguments{
\item{loss}{L1, L2, L3 and L4 representing the loss function used.}

\item{a1}{Hyperparameter of the prior distribution for beta.}

\item{b1}{Hyperparameter of the prior distribution for beta.}

\item{a2}{Hyperparameter of the prior distribution for alpha^2.}

\item{b2}{Hyperparameter of the prior distribution for alpha^2.}

\item{c}{A positive real number representing the cost of colect one aliquot.}

\item{rho}{A number in (0, 1). The probability of the credible interval is $1-rho$. Only
for lost function 1.}

\item{gam}{A positive real number connected with the credible interval when using lost
function 2.}

\item{nmax}{A positive integer representing the maximum number for compute the Bayes risk.
Default is 100.}

\item{nlag}{A positive integer representing the lag in the n's used to compute the Bayes risk. Default is 10.}

\item{nrep}{A positive integer representing the number of samples taken for each $n$.}

\item{lrep}{A positive integer representing the number of samples taken for $S_n$. Default is 100.}

\item{npost}{A positive integer representing the number of values to draw from the posterior distribution of the mean. Default is 100.}

\item{plot}{Boolean. If TRUE (default) it plot the estimated Bayes risks and the fitted curve.}

\item{...}{Currently ignored.}
}
\value{
An integer representing the sample size.
}
\description{
A function
}
