% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCodes.R
\name{logp.beta}
\alias{logp.beta}
\title{x}
\usage{
logp.beta(beta, x, a1, b1, a2, b2)
}
\arguments{
\item{beta}{scale parameter.}

\item{x}{vector of observed values of the model.}

\item{a1}{hyperparameter of the prior distribution for beta.}

\item{b1}{hyperparameter of the prior distribution for beta.}

\item{a2}{hyperparameter of the prior distribution for \code{shape^2}.}

\item{b2}{hyperparameter of the prior distribution for \code{shape^2}.}
}
\value{
The log da marginal posterior of beta
}
\description{
x
}
