% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCodes.R
\name{rbeta.post}
\alias{rbeta.post}
\title{Random generation for the joint posterior distribution of the Birnbaum-Saunders/inverse-gamma model.}
\usage{
rbeta.post(N, x, a1, b1, a2, b2, burnin, thin, start, varcov, scale = 1)
}
\arguments{
\item{N}{number of observations.}

\item{x}{vector of observed values of the model.}

\item{a1}{hyperparameter of the prior distribution for beta.}

\item{b1}{hyperparameter of the prior distribution for beta.}

\item{a2}{hyperparameter of the prior distribution for \code{shape^2}.}

\item{b2}{hyperparameter of the prior distribution for \code{shape^2}.}

\item{burnin}{a positive constant for the sampling method.}

\item{thin}{a positive constant for the sampling method.}

\item{start}{a positive constant for the sampling method.}

\item{varcov}{a positive constant for the sampling method.}

\item{scale}{a positive constant for the sampling method.}
}
\value{
A random sample of the joint posterior distribution of the model Birnbaum-Saunders/inverse-gamma model.
}
\description{
A function for generating values from the joint posterior distribution of the Birnbaum-Saunders/inverse-gamma model.
}
\references{
Wang, M., Sun, X. and Park, C. (2016) Bayesian analysis of Birnbaum-Saunders distribution via the generalized ratio-of-uniforms methods. Comput. Stat. 31: 207--225.
}
\author{
Eliardo G. Costa \email{eliardocosta@ccet.ufrn.br} and Manoel Santos-Neto \email{manoel.ferreira@ufcg.edu.br}
}
