% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCodes.R
\docType{data}
\name{repair}
\alias{repair}
\title{Repair times}
\format{
A data frame with 46 observations:
\describe{
  \item{repair}{repair times}
}
}
\usage{
repair
}
\description{
These observations correspond to maintenance data on active repair times (in hours) for an airborne communications transceiver.
}
\references{
Hsieh, H.K. Estimating the critical time of the inverse Gaussian hazard rate.  \emph{IEEE Trans. Reliab.} 1990, \emph{39}, 342--345.
}
\keyword{datasets}
