% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCodes.R
\name{rpost.bs}
\alias{rpost.bs}
\title{Random generation for the posterior distribution of the Birnbaum-Saunders/inverse-gamma model.}
\usage{
rpost.bs(N, x, a1, b1, a2, b2, r)
}
\arguments{
\item{N}{Number of observations.}

\item{x}{Observed values of the model.}

\item{a1}{Hyperparameter of the prior distribution for beta.}

\item{b1}{Hyperparameter of the prior distribution for beta.}

\item{a2}{Hyperparameter of the prior distribution for alpha^2.}

\item{b2}{Hyperparameter of the prior distribution for alpha^2.}

\item{r}{A constant for the sampling method.}
}
\value{
A random sample of the posterior distribution of the model Birnbaum-Saunders/inverse-gamma.
}
\description{
The function
}
\note{
We adapted the available R script (Supplementary Material) of Wang, Sun & Park (2016, Comp. Stat.)
}
