% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCodes.R
\name{rbs}
\alias{rbs}
\title{Random generation for the Birnbaum-Saunders distribution.}
\usage{
rbs(n, alpha, beta)
}
\arguments{
\item{n}{Sample size.}

\item{alpha}{Shape parameter.}

\item{beta}{Scale parameter.}
}
\value{
A sample of size n from the Birnbaum-Saunders distribution.
}
\description{
The random generation for the Birnbaum-Saunders distribution.
}
