% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCodes.R
\name{rbs}
\alias{rbs}
\title{Random generation for the Birnbaum-Saunders distribution.}
\usage{
rbs(n=1.0, alpha = 0.5, beta = 1.0)
}
\arguments{
\item{n}{sample size.}

\item{alpha}{shape parameter.}

\item{beta}{scale parameter.}
}
\value{
A sample of size n from the Birnbaum-Saunders distribution.
}
\description{
The random generation for the Birnbaum-Saunders distribution.
}
